requirements.txt
anyio==3.6.1
apache-libcloud==3.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
fastapi==0.78.0
h11==0.13.0
idna==3.3
mpmath==1.2.1
pycparser==2.21
pydantic==1.9.1
python-dotenv==0.20.0
requests==2.28.0
sniffio==1.2.0
starlette==0.19.1
sympy==1.10.1
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn==0.18.1
